requirements.txt
streamlit
yfinance
google-generativeai
pytz